Fix too much text being deleted on removing a binding
If the ',' before the binding section being removed is the first
character on the line the wrong lines were being removed. The
GetExtentInfo was stopping on the line before so the startIndex
was being considered on the wrong line so you would end up with two
lines being selected as the extent instead of the single line.